requirements.txt
flask
pillow
torch
torchvision
pandas
requests
opencv-python
tqdm
pyyaml
matplotlib
seaborn
scipy